Feat(dbViewer): show OTP codes in Group Viewer (closes #8)
Revealed on button tap.

+++ KeePassium/database/entry/ViewableEntryFieldFactory.swift
@@ -109,7 +109,7 @@ class TOTPViewableField: DynamicViewableField {
         return value
     }
     
-    var elapsedTimeFraction: Float? {
+    var elapsedTimeFraction: Double? {
         return totpGenerator?.elapsedTimeFraction
     }
